(refactor): create validators dir
